+++ clang-tools-extra/clang-tidy/hooks/VerifyBufLenCheck.h
@@ -1,0 +1,37 @@
+//===--- VerifyBufLenCheck.h - clang-tidy -----------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_VERIFYBUFLENCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_VERIFYBUFLENCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+/// Verifying a cryptographic signature by calling util_verify
+/// requires valid public key & data signature.
+class VerifyBufLenCheck : public ClangTidyCheck {
+public:
+  VerifyBufLenCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  const int64_t KEY_SIZE = 33; // value copied from rippled source
+  const int64_t MIN_SIG_SIZE = 8; // value copied from rippled source
+  const int64_t MAX_SIG_SIZE = 72; // value copied from rippled source
+};
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_VERIFYBUFLENCHECK_H